[IMP] role_policy - add 'roles' attribute to views
In order to compensate for the capability to add groups on view elements a new
attribute has been added: "roles" with attritue value a list of role codes.

The roles attribute works similar as the groups attribute:
the view element will be removed if a user does not belong to one of the specified roles.