[FIX]role_policy_hr : fix hr_employee access
An employee record is stored in the database via the hr_employee table
but only a subset of these fields are available to the regular users.
Only users belonging to the "hr.group_hr_user" can retrieve all fields.
This commit adds "hr.group_hr_user" to the role policy untouchable groups.